requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.1

# Database (CORREÇÃO: versão estável do mysql-connector)
PyMySQL==1.1.2
mysql-connector-python==8.2.0

# Environment Configuration
python-dotenv==1.2.1

# HTTP & Requests
requests==2.31.0

# File Processing
openpyxl==3.1.5
xlsxwriter==3.2.9

# Web Server (IMPORTANTE para Render)
gunicorn==21.2.0

# Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1